init: add settings
