+++ nodemon.json
@@ -1,0 +1,5 @@
+{
+  "execMap": {
+      "ts": "node --loader ts-node/esm"
+  }
+}
